fix: arreglar metodos  para  que  devuelvan errores

+++ Tareas.API/Services/Implementaciones/TareaService.cs
@@ -28,7 +28,8 @@ namespace Tareas.API.Services.Implementaciones
         /// <returns>Un resultado indicando si se completó la tarea y la tarea</returns>
         public async Task<ServiceResponse<Tarea>> CrearTareaAsync(CrearTareaDTO tarea)
         {
-            ValidacionesTarea.EsTareaValida(tarea);
+            if (!ValidacionesTarea.EsTareaValida(tarea).Success)
+                return ServiceResponse<Tarea>.Error("Los datos de la tarea son inválidos.");
 
             var tareaACrear = TareaMapper.ToModel(tarea);
 
@@ -44,11 +45,13 @@ namespace Tareas.API.Services.Implementaciones
         /// <returns>Respuesta con la tarea obtenida y mensaje de éxito o error.</returns
         public async Task<ServiceResponse<Tarea>> ObtenerTareaPorIdAsync(string id)
         {
-            ValidacionesTarea.EsIdValido(id);
+            if (!ValidacionesTarea.EsIdValido(id).Success)
+                return ServiceResponse<Tarea>.Error("El ID de la tarea es inválido.");
 
             var tarea = await _context.GetByIdAsync(id);
 
-            ValidacionesTarea.EsTareaExistente(id, tarea);
+            if (!ValidacionesTarea.EsTareaExistente(id, tarea).Success)
+                return ServiceResponse<Tarea>.Error("La tarea no existe.");
 
             return ServiceResponse<Tarea>.Ok(tarea, "Tarea obtenida exitosamente");
         }
@@ -60,13 +63,16 @@ namespace Tareas.API.Services.Implementaciones
         /// <param name="tarea">DTO con los nuevos datos de la tarea y un mensaje  de exito o error</param
         public async Task<ServiceResponse<Tarea>> ActualizarTareaAsync(string id, ActualizarTareaDTO tarea)
         {
-            ValidacionesTarea.EsIdValido(id);
+            if (!ValidacionesTarea.EsIdValido(id).Success)
+                return ServiceResponse<Tarea>.Error("El ID de la tarea es inválido.");
 
-            ValidacionesTarea.EsTareaActualizable(id, tarea);
+            if (!ValidacionesTarea.EsTareaActualizable(id, tarea).Success)
+                return ServiceResponse<Tarea>.Error("Los datos de la tarea son inválidos.");
 
             var existingTarea = await _context.GetByIdAsync(id);
 
-            ValidacionesTarea.EsTareaExistente(id, existingTarea);
+            if (!ValidacionesTarea.EsTareaExistente(id, existingTarea).Success)
+                return ServiceResponse<Tarea>.Error("La tarea no existe.");
 
             await _context.UpdateAsync(id, TareaMapper.ToModel(existingTarea, tarea));
 
@@ -80,11 +86,13 @@ namespace Tareas.API.Services.Implementaciones
         /// <returns>Respuesta con la tarea eliminada y mensaje de éxito o error.</returns
         public async Task<ServiceResponse<Tarea>> EliminarTareaAsync(string id)
         {
-            ValidacionesTarea.EsIdValido(id);
+            if  (!ValidacionesTarea.EsIdValido(id).Success)
+                return ServiceResponse<Tarea>.Error("El ID de la tarea es inválido.");
 
             var tarea = await _context.GetByIdAsync(id);
 
-            ValidacionesTarea.EsTareaExistente(id, tarea);
+            if (!ValidacionesTarea.EsTareaExistente(id, tarea).Success)
+                return ServiceResponse<Tarea>.Error("La tarea no existe.");
 
             await _context.DeleteAsync(id);
